zzogl-pg: Work on reducing the amount of platform specific code a bit.
git-svn-id: http://pcsx2.googlecode.com/svn/trunk@2927 96395faa-99c1-11dd-bbfe-3dabce05a288

+++ plugins/zzogl-pg/opengl/GS.h
@@ -444,7 +444,7 @@ union tex_0_info
 		if (tbw == 0)
 			return 64;
 		else
-			return (tbw << 6);
+			return ((u32)tbw << 6);
 	}
 	u32 psm_fix()
 	{
